Our JS combiner translates
var fixtureName = (undefined !== this.activeFixture) ? ...

to

var fixtureName = undefined !== this.activeFixture ? ...

and jshint then thinks that the resulting JS is not valid, because it
doesn't like "fixtureName = undefined". Stupid jshint gives the
following error:

"ext/imvutest.async.js: line 61, col 37, It's not necessary to initialize
'fixtureName' to 'undefined'."

Thus, I figured that I'll just swap the variable and the "undefined".