Fix loading issue with multiple imvu.js included
If imvu.min.js was included after an existing imvujs was already on the
page, module resolution would fail. If imvu.js was included after an
existing imvujs was already on the page, a harmless error was thrown.

How did this happen?

1. imvu.js defines window.module (A.module), which has private state
   'currentModuleURL' initially set to undefined
2. module.run() gets called, which modifies 'currentModuleURL' private
   state
3. imvu.js redefines window.module (B.module)
4. XMLHttpRequest for the dependency comes back to A.module's handler,
   which evaluates JS in the global context
5. Evaluated javascript calls window.module, which resolves to B.module
6. B.module checks private state of 'currentModuleURL' and sees that it
   is undefined, so it throws an error, assuming that it did not
   originate from a module.run call. B.module throws and the module
   contents never gets executed.

Why did unminified imvujs produce a harmless error?

Esprima is only included in the unminified imvujs, which attempts to
call define() if it exists. define() is a thin wrapper around module(),
which shouldn't be called through a <script> tag, so the script errors
before being successfully loaded.