deferred custom loaders for the combiner
An action is a string that contains a bang "!", which separates
arguments. There must be at least two arguments.

The first argument is *always* the name of the custom loader. The second
argument is *always* a dependency path, which is relative to the
dependency itself.

There currently is no consumer of these from node-module.js or
web-module.js.

Did I mention how much I like typescript?